+++ polaris-cli/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @shopify/polaris-cli
 
+## 0.2.8
+
+## 0.2.7
+
+## 0.2.6
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @shopify/polaris-migrator@0.19.3
+
 ## 0.2.5
 
 ### Patch Changes